Added Marvin1 and Marvin4
Marvin1 is the beginning of me programming the Marvin program, I call it Fake-GPT since you can chat with it but it is not a real GPT :). The fourth version is the final version and the versions in between I did not include as they just add more features that Marvin4 already has.